requirements.txt
requests>=2.31.0
pandas>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
openai>=1.10.0
python-dotenv>=1.0.0
